+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## New Features
 - PR #287 SNMG power iteration step1 
 - PR #297 SNMG degree calculation 
+- PR #300 Personalized Page Rank
 - PR #302 SNMG CSR Pagerank (cuda/C++)
+- PR #315 Weakly Connected Components adapted from cuML (cuda/C++)
 
 ## Improvements
 - PR #291 nvGraph is updated to use RMM instead of directly invoking cnmem functions.
@@ -19,8 +21,9 @@
 - PR #306 Fixed cmake warnings due to library conflicts.
 - PR #311 Fixed bug in SNMG degree causing failure for three gpus
 - PR #309 Update conda build recipes
+- PR #314 Added datasets to gitignore
 
-# cuGraph 0.7.0 (Date TBD)
+# cuGraph 0.7.0 (10 May 2019)
 
 ## New Features
 - PR #195 Added Graph.get_two_hop_neighbors() method